Small changes in SysLog server (Component and demo)
OverbyteIcsSysLogServerDemo:
FSysLogServer component is created runtime on form creation, but 
SysLogServer1DataAvailable method is not assigned to 
FSysLogServer.OnDataAvailable anywhere.

TSysLogServer.ParseRawMessageRFC5424:
 From line 400 to 457 Date and Time values are read from message 
respectively to SYear, SMonth,SDay,SHour,SMinute,SSecond and SMSec
but if there is a Time Bias, SHour,SMinute are owerritten is case 
construct (Line 463,471)
After case construct in line 493,494 DecodedMessage.Hour and 
DecodedMessage.Minute will get values from the overwritten variables.

If client hostname contains negative sign "-" , for example 
"somename-peter", than between lines 520-532 it is evaluated as "invalid 
PID delimter" and an exception is raised.
I dont know if this is a valid hostname for that RFC

Regards
Peter